requirements.txt
django==3.1.4
djangorestframework==3.12.4
mysqlclient==2.0.3
pytest-django==4.2.0
Pillow==8.2.0